backend/requirements.txt
flask
flask-cors
flask-bcrypt
pyjwt
joblib
pandas
numpy
python-dotenv
